requirements.txt
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.50.0
requests==2.31.0  # Required for making HTTP requests
